8845 FIX Web-API: Prevent adding wrong encoded labels to hosts / rulesets
In previous 1.6 releases the Web API did not validate the labels that
could e.g. be added to hosts strictly enough. When using "python" as
"request_format" this could result in a broken configuration, leading
to exceptions in several places of the GUI.

The Web API now ensures the received labes are all unicode strings as
Checkmk expects.

Change-Id: I1bcb8b8411474a7e7fa6a4a3a281fbe4ebfb850e