8901 FIX Fixed calculation of latency for netapp check
Average latency, average read latency, and average write latency
are now calculated as a function of total_ops instead of the
time.  This complies with the formula from NetApp's Performance
Management Design Guide.

According to this guide, latency should be calculated as:

           latency(now) - latency(before)
    lat = ---------------------------------
          total_ops(now) - total_ops(before)

and not as a function of time.

This fixes FEED-3874.

Change-Id: I527e8e77cff016559ae85e6bed173655bc34dfb4