8902 apache_status: Set optional instance name
Before, the status was reported as
  Apache localhost $port Status
or
  Apache $fqdn $port Status

Now, if an instance name is provided, the report is
  Apache $instance $port Status
if $instance is provided.  Otherwise, the reporting uses hostname/fqdn
as before.

Bakery
======

The GUI has a new "instance" element.

The bakelet does not need to be modified.

Agent and check
===============

The separator is set to | to allow spaces in the instance name.  This
implies rewriting a relatively large part of the parser to accomodate
the new format.

Further changes to the parser are:
 - EAFP
 - defaultdict (the check is about 30% faster and it is simpler without
 the explicit `__contains__()` check)

This is part 2 of CMK-1853.

Change-Id: I4efc2608a448dde51c9d02286a464eb34ba0631a