8866 FIX Docker: Fix GUI redirects (1.6 regression)
When directly using the GUI of the official Checkmk docker container without a reverse proxy
in front of it, the redirects of URLs like <tt>http://[address]/cmk</tt> were not working in
the previous 1.6 releases.

The redirects have now been fixed. In addition we have now added a new redirect from the top
level URL <tt>http://[address]/</tt> which redirects the user to the GUI.

Change-Id: Id2928eb0aed064da0ee390d4aed8e7c1a3c8f868