Resolve scrolling error in the Chrome console
When a user scrolled inside the sidebar errors with the message:

Unable to preventDefault inside passive event listener due to target
being treated as passive.

were logged to the console. This is fixed by removing the
preventDefault call. Since the sidebar always has a height of
100% no default scrolling action will be applied. This was the
behaviour for Chrome anyway since the event listener is passive by
default:

https://www.chromestatus.com/features/6662647093133312

Change-Id: I4ab522ad4e4e2b2b2221397817296fde8a95fce4