8889 FIX SLES15: Fix building MSI packages
The MSI utilities, which should be shipped with Checkmk, were missing in
previous 1.6 packages.

msitools depends on libgsf, but our own libgsf package was not used for
building the msitools package.

The libgsf is currently only built for SLES15, so we can only base
on this library on this distribution.

Could be a good next step to build libgsf on all distributions and
make the msitools always base on our self built library. We could
then remove another runtime dependency.

CMK-2627

Change-Id: I5183664023ae054bcf69f1d99291a03f16149d7f