8996 check_icmp: minimum number of successful pings for OK status consideration
In the previous version, failure to ping one IP address amongst
multiples addresses led to a Crit status of the icmp service. This new
feature allows to specify the minimum number of successful pings in
order to deem the service as OK. This signifies that pinging one address
successfully can be sufficient regardless of the number of failures for
the other remanining addresses.

Change-Id: I83c7738ff2c7dd74a878360b023982f16a2520f9