7171 FIX cmk-update-agent: Improved error handling
Previously, the agent updater might crash with a meaningless error
like <pre>ERROR compile() expected string without null bytes</pre>
when it tries to read from a corrupt config or state file.

This is now mitigated by showing a meaningful error that refers to
the file that caused the problem.

Jira: CMK-2634

Change-Id: I114b112fe25caaa1d61933940032279ecc53a874