10034 FIX Fix wrong encoding in site user context
At least since werk #7330 the environment variables LANG and LC_ALL were set to
"C" to have a localization independent runtime environment in the sites. This
lead to encoding problems in shell context, which also inherited into e.g.
python runtimes, changing the preferred encoding of the locale module.

To fix this issue the site is now setting "C.UTF-8" as environment.

CMK-2615

Change-Id: If6681b2f12b64b91b730873e683fbca7bc12d1bf