8914 FIX Fix calculation of latency for netapp_api_vs_traffic 2
read_ops and write_ops do not necessarily exist to use as
reference for the avg_read_latency and avg_write_latency.
The check now crashes when they are not available.

This fix uses the latency and ops values in the isci and fcp
namespaces and handles missing references.  When the ref
is missing, the time is taken as reference.

This is the second try at FEED-3874 (api_vs_traffic).

Change-Id: I0084d74dd613ec68e33c34e8ef28f78a2baef637