10188 Windows Agent reinstalls files using also content of those files
Previously Windows Agent used only the file timestamp to determine whether
the file have to be reinstalled.
This may lead to impossibility to update the file if for some reason the
file timestamp was wrong(downgrade, invalid time).

Now Windows Agent compares file content too. If content mismatches,
then the file should be installed even if timestamp is ok.

Change-Id: Ifa2fe080ef1510845a8c139955de8e2be7089307