10246 FIX omd backup: Made it more robust against rrdcached communication issues
The "omd backup" could be terminated during backup by issues of the rrdcached.
The backup communicates with the rrdcached to pause updates to the RRD files
to ensure the files are consistent during backup. In case the rrdcached is not
answering as expected or a connection is not possible the backup should
continue with the backup instead of failing.

Change-Id: I00f79c63c3992e505cf7f372e3ca479388e1aad6